requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
jinja2==3.1.6
pyyaml==6.0.2
httpx==0.28.1
pytest==8.4.2
